svelte-migrate: renamed files
