+++ src/routes/api/locale/+server.ts
@@ -21,6 +21,9 @@ export async function GET(event: RequestEvent) {
   // TODO: Reworking typing here to index the locales object
   const locale = getLocale(userLocale, '_') as 'ca_en' | 'ca_fr' | 'us_en' | 'us_es';
 
+  throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+  // Suggestion (check for correctness before using):
+  // return new Response(locales[locale]);
   return {
     body: locales[locale],
   };
