--- src/routes/callback.ts
@@ -1,6 +1,0 @@
-export async function GET() {
-  return {
-    status: 200,
-    body: 'OK',
-  };
-}
