+++ src/routes/callback/+server.ts
@@ -1,0 +1,3 @@
+export async function GET() {
+  return new Response('OK');
+}
