+++ src/routes/e2e/widget/+layout@clean.svelte
