+++ src/routes/e2e/widget/modal/+page.svelte
