+++ src/routes/proxy/authenticate/+server.ts
@@ -1,3 +1,4 @@
+import { json as json$1 } from '@sveltejs/kit';
 import type { RequestEvent } from '@sveltejs/kit';
 
 import { AM_COOKIE_NAME, AM_DOMAIN_PATH, JSON_REALM_PATH } from '$lib/constants';
@@ -47,6 +48,15 @@ export async function POST(event: RequestEvent) {
     }
   }
 
+  throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+  // Suggestion (check for correctness before using):
+  // return json$1(resBody, {
+  //   headers: {
+  //     'set-cookie': cookieUuid
+  //       ? `cookie=${cookieUuid}; domain=.crbrl.ngrok.io; SameSite=None; HTTPOnly; Secure;`
+  //       : '',
+  //   }
+  // });
   return {
     status: 200,
     body: resBody,
