+++ src/routes/proxy/sessions/+server.ts
@@ -19,6 +19,7 @@ export async function POST(event: RequestEvent) {
   // const resBody = await response.json();
   // console.log(response);
 
+  throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
   return {
     status: 200,
     body: response.body,
