+++ src/routes/register/+page.svelte
@@ -1,24 +1,28 @@
 <script context="module" lang="ts">
-  import Box from '$components/primitives/box/centered.svelte';
-  // import { initialize, type InitObject } from '$journey/journey.store';
+  throw new Error("@migration task: Check code was safely removed (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292722)");
 
-  import '../app.css';
+  // import Box from '$components/primitives/box/centered.svelte';
+  // // import { initialize, type InitObject } from '$journey/journey.store';
 
-  // Loads first step on server.
-  // IMPORTANT: Requires `window` to be removed from timeout utility in JS SDK
-  // export async function load() {
-  //   let initObj = await initialize('Registration');
+  // import '../../app.css';
 
-  //   return {
-  //     props: { initObj },
-  //   }
-  // }
+  // // Loads first step on server.
+  // // IMPORTANT: Requires `window` to be removed from timeout utility in JS SDK
+  // // export async function load() {
+  // //   let initObj = await initialize('Registration');
+
+  // //   return {
+  // //     props: { initObj },
+  // //   }
+  // // }
 </script>
 
 <script lang="ts">
+  throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
   import { onMount } from 'svelte';
 
-  import Journey from '../lib/journey/journey.svelte';
+  import Journey from '../../lib/journey/journey.svelte';
   import { initialize as initializeJourney, type JourneyStore } from '$journey/journey.store';
   import { initialize as initializeContent } from '$lib/locale.store';
   import { initialize as initializeOAuth, type OAuthStore } from '$lib/oauth/oauth.store';
