+++ src/routes/register/+page.ts
@@ -1,0 +1,14 @@
+import Box from '$components/primitives/box/centered.svelte';
+// import { initialize, type InitObject } from '$journey/journey.store';
+
+import '../../app.css';
+
+// Loads first step on server.
+// IMPORTANT: Requires `window` to be removed from timeout utility in JS SDK
+// export async function load() {
+//   let initObj = await initialize('Registration');
+
+//   return {
+//     props: { initObj },
+//   }
+// }
